refactored test_resources.py into test_resources package (#1672)
* refactored test_resources.py into test_resources package

* Removed all .DS_Store files

* updated the changelog

* updated target string
