+++ AUTHORS
@@ -146,3 +146,4 @@ The following is a list of much appreciated contributors:
 * EricOuma (Eric Ouma)
 * Glay00 (Gleb Gorelov)
 * PrashansaChaudhary (Prashansa Chaudhary)
+* Vedang Barhate (bvedang)
