+++ docs/changelog.rst
@@ -18,6 +18,7 @@ Please refer to :doc:`release notes<release_notes>`.
 - Standardised interface of :meth:`~import_export.widgets.Widget.render` (#1657)
 - Improved documentation relating to validation on import (#1665)
 - Refactored test_admin_integration: split into smaller test modules (#1662)
+- Refactored test_resources: split into smaller test modules (#1672)
 
 4.0.0-alpha.5 (2023-09-22)
 --------------------------
