+++ tests/core/tests/test_resources/test_resources.py
@@ -8,6 +8,27 @@ from unittest import mock, skipUnless
 from unittest.mock import patch
 
 import tablib
+from core.models import (
+    Author,
+    Book,
+    Category,
+    Entry,
+    Profile,
+    WithDynamicDefault,
+    WithFloatField,
+)
+from core.tests.resources import (
+    AuthorResource,
+    AuthorResourceWithCustomWidget,
+    BookResource,
+    BookResourceWithLineNumberLogger,
+    BookResourceWithStoreInstance,
+    CategoryResource,
+    MyResource,
+    ProfileResource,
+    WithDefaultResource,
+)
+from core.tests.utils import ignore_widget_deprecation_warning
 from django.conf import settings
 from django.contrib.auth.models import User
 from django.core.exceptions import (
@@ -26,35 +47,6 @@ from django.utils.html import strip_tags
 from import_export import exceptions, fields, resources, results, widgets
 from import_export.instance_loaders import ModelInstanceLoader
 from import_export.resources import Diff
-
-from ..models import (
-    Author,
-    Book,
-    Category,
-    Entry,
-    Person,
-    Profile,
-    Role,
-    UUIDBook,
-    UUIDCategory,
-    WithDefault,
-    WithDynamicDefault,
-    WithFloatField,
-)
-from .utils import ignore_widget_deprecation_warning
-
-
-class MyResource(resources.Resource):
-    name = fields.Field()
-    email = fields.Field()
-    extra = fields.Field()
-
-    def __init__(self, **kwargs):
-        super().__init__(**kwargs)
-        self.kwargs = kwargs
-
-    class Meta:
-        export_order = ("email", "name")
 
 
 class ResourceTestCase(TestCase):
@@ -170,7 +162,7 @@ class ResourceTestCase(TestCase):
     def test_get_field_name_raises_AttributeError(self):
         err = (
             "Field x does not exists in <class "
-            "'core.tests.test_resources.MyResource'> resource"
+            "'core.tests.resources.MyResource'> resource"
         )
         with self.assertRaisesRegex(AttributeError, err):
             self.my_resource.get_field_name("x")
@@ -194,90 +186,6 @@ class ResourceTestCase(TestCase):
         full_clean_mock.assert_called_once_with(
             exclude=target.keys(), validate_unique=True
         )
-
-
-class AuthorResource(resources.ModelResource):
-    books = fields.Field(
-        column_name="books",
-        attribute="book_set",
-        readonly=True,
-    )
-
-    class Meta:
-        model = Author
-        export_order = ("name", "books")
-
-
-class BookResource(resources.ModelResource):
-    published = fields.Field(column_name="published_date")
-
-    class Meta:
-        model = Book
-        exclude = ("imported",)
-
-
-class BookResourceWithStoreInstance(resources.ModelResource):
-    class Meta:
-        model = Book
-        store_instance = True
-
-
-class BookResourceWithLineNumberLogger(BookResource):
-    def __init__(self, *args, **kwargs):
-        self.before_lines = []
-        self.after_lines = []
-        return super().__init__(*args, **kwargs)
-
-    def before_import_row(self, row, **kwargs):
-        row_number = kwargs.pop("row_number")
-        self.before_lines.append(row_number)
-
-    def after_import_row(self, row, row_result, **kwargs):
-        row_number = kwargs.pop("row_number")
-        self.after_lines.append(row_number)
-
-
-class CategoryResource(resources.ModelResource):
-    class Meta:
-        model = Category
-
-
-class ProfileResource(resources.ModelResource):
-    class Meta:
-        model = Profile
-        exclude = ("user",)
-
-
-class WithDefaultResource(resources.ModelResource):
-    class Meta:
-        model = WithDefault
-        fields = ("name",)
-
-
-class HarshRussianWidget(widgets.CharWidget):
-    def clean(self, value, row=None, *args, **kwargs):
-        raise ValueError("Ова вриједност је страшна!")
-
-
-class AuthorResourceWithCustomWidget(resources.ModelResource):
-    class Meta:
-        model = Author
-
-    @classmethod
-    def widget_from_django_field(cls, f, default=widgets.Widget):
-        if f.name == "name":
-            return HarshRussianWidget
-        result = default
-        internal_type = (
-            f.get_internal_type()
-            if callable(getattr(f, "get_internal_type", None))
-            else ""
-        )
-        if internal_type in cls.WIDGETS_MAP:
-            result = cls.WIDGETS_MAP[internal_type]
-            if isinstance(result, str):
-                result = getattr(cls, result)(f)
-        return result
 
 
 class ModelResourcePostgresModuleLoadTest(TestCase):
@@ -1811,1240 +1719,3 @@ if "postgresql" in settings.DATABASES["default"]["ENGINE"]:
 
             self.book.refresh_from_db()
             self.assertEqual(self.book.data, self.data)
-
-
-class ForeignKeyWidgetFollowRelationship(TestCase):
-    def setUp(self):
-        self.user = User.objects.create(username="foo")
-        self.role = Role.objects.create(user=self.user)
-        self.person = Person.objects.create(role=self.role)
-
-    @ignore_widget_deprecation_warning
-    def test_export(self):
-        class MyPersonResource(resources.ModelResource):
-            role = fields.Field(
-                column_name="role",
-                attribute="role",
-                widget=widgets.ForeignKeyWidget(Role, field="user__username"),
-            )
-
-            class Meta:
-                model = Person
-                fields = ["id", "role"]
-
-        resource = MyPersonResource()
-        dataset = resource.export(Person.objects.all())
-        self.assertEqual(len(dataset), 1)
-        self.assertEqual("1", dataset[0][0])
-        self.assertEqual("foo", dataset[0][1])
-
-        self.role.user = None
-        self.role.save()
-
-        resource = MyPersonResource()
-        dataset = resource.export(Person.objects.all())
-        self.assertEqual(len(dataset), 1)
-        self.assertEqual("1", dataset[0][0])
-        self.assertEqual(None, dataset[0][1])
-
-
-class ManyRelatedManagerDiffTest(TestCase):
-    fixtures = ["category", "book", "author"]
-
-    def setUp(self):
-        pass
-
-    @ignore_widget_deprecation_warning
-    def test_related_manager_diff(self):
-        dataset_headers = ["id", "name", "categories"]
-        dataset_row = ["1", "Test Book", "1"]
-        original_dataset = tablib.Dataset(headers=dataset_headers)
-        original_dataset.append(dataset_row)
-        dataset_row[2] = "2"
-        changed_dataset = tablib.Dataset(headers=dataset_headers)
-        changed_dataset.append(dataset_row)
-
-        book_resource = BookResource()
-        export_headers = book_resource.get_export_headers()
-
-        add_result = book_resource.import_data(original_dataset, dry_run=False)
-        expected_value = '<ins style="background:#e6ffe6;">1</ins>'
-        self.check_value(add_result, export_headers, expected_value)
-        change_result = book_resource.import_data(changed_dataset, dry_run=False)
-        expected_value = (
-            '<del style="background:#ffe6e6;">1</del>'
-            '<ins style="background:#e6ffe6;">2</ins>'
-        )
-        self.check_value(change_result, export_headers, expected_value)
-
-    def check_value(self, result, export_headers, expected_value):
-        self.assertEqual(len(result.rows), 1)
-        diff = result.rows[0].diff
-        self.assertEqual(diff[export_headers.index("categories")], expected_value)
-
-
-class ManyToManyWidgetDiffTest(TestCase):
-    # issue #1270 - ensure ManyToMany fields are correctly checked for
-    # changes when skip_unchanged=True
-    fixtures = ["category", "book", "author"]
-
-    def setUp(self):
-        pass
-
-    @ignore_widget_deprecation_warning
-    def test_many_to_many_widget_create(self):
-        # the book is associated with 0 categories
-        # when we import a book with category 1, the book
-        # should be updated, not skipped
-        book = Book.objects.first()
-        book.categories.clear()
-        dataset_headers = ["id", "name", "categories"]
-        dataset_row = [book.id, book.name, "1"]
-        dataset = tablib.Dataset(headers=dataset_headers)
-        dataset.append(dataset_row)
-
-        book_resource = BookResource()
-        book_resource._meta.skip_unchanged = True
-        self.assertEqual(0, book.categories.count())
-
-        result = book_resource.import_data(dataset, dry_run=False)
-
-        book.refresh_from_db()
-        self.assertEqual(1, book.categories.count())
-        self.assertEqual(
-            result.rows[0].import_type, results.RowResult.IMPORT_TYPE_UPDATE
-        )
-        self.assertEqual(Category.objects.first(), book.categories.first())
-
-    @ignore_widget_deprecation_warning
-    def test_many_to_many_widget_create_with_m2m_being_compared(self):
-        # issue 1558 - when the object is a new instance and m2m is
-        # evaluated for differences
-        dataset_headers = ["id", "categories"]
-        dataset_row = ["1", "1"]
-        dataset = tablib.Dataset(headers=dataset_headers)
-        dataset.append(dataset_row)
-        book_resource = BookResource()
-        book_resource._meta.skip_unchanged = True
-
-        result = book_resource.import_data(dataset, dry_run=False)
-
-        self.assertFalse(result.has_errors())
-        self.assertEqual(len(result.rows), 1)
-        self.assertEqual(result.rows[0].import_type, results.RowResult.IMPORT_TYPE_NEW)
-
-    @ignore_widget_deprecation_warning
-    def test_many_to_many_widget_update(self):
-        # the book is associated with 1 category ('Category 2')
-        # when we import a book with category 1, the book
-        # should be updated, not skipped, so that Category 2 is replaced by Category 1
-        book = Book.objects.first()
-        dataset_headers = ["id", "name", "categories"]
-        dataset_row = [book.id, book.name, "1"]
-        dataset = tablib.Dataset(headers=dataset_headers)
-        dataset.append(dataset_row)
-
-        book_resource = BookResource()
-        book_resource._meta.skip_unchanged = True
-        self.assertEqual(1, book.categories.count())
-
-        result = book_resource.import_data(dataset, dry_run=False)
-        self.assertEqual(
-            result.rows[0].import_type, results.RowResult.IMPORT_TYPE_UPDATE
-        )
-        self.assertEqual(1, book.categories.count())
-        self.assertEqual(Category.objects.first(), book.categories.first())
-
-    @ignore_widget_deprecation_warning
-    def test_many_to_many_widget_no_changes(self):
-        # the book is associated with 1 category ('Category 2')
-        # when we import a row with a book with category 1, the book
-        # should be skipped, because there is no change
-        book = Book.objects.first()
-        dataset_headers = ["id", "name", "categories"]
-        dataset_row = [book.id, book.name, book.categories.first().id]
-        dataset = tablib.Dataset(headers=dataset_headers)
-        dataset.append(dataset_row)
-
-        book_resource = BookResource()
-        book_resource._meta.skip_unchanged = True
-
-        self.assertEqual(1, book.categories.count())
-        result = book_resource.import_data(dataset, dry_run=False)
-        self.assertEqual(result.rows[0].import_type, results.RowResult.IMPORT_TYPE_SKIP)
-        self.assertEqual(1, book.categories.count())
-
-    @ignore_widget_deprecation_warning
-    def test_many_to_many_widget_handles_ordering(self):
-        # the book is associated with 2 categories ('Category 1', 'Category 2')
-        # when we import a row with a book with both categories (in any order), the book
-        # should be skipped, because there is no change
-        book = Book.objects.first()
-        self.assertEqual(1, book.categories.count())
-        cat1 = Category.objects.get(name="Category 1")
-        cat2 = Category.objects.get(name="Category 2")
-        book.categories.add(cat1)
-        book.save()
-        self.assertEqual(2, book.categories.count())
-        dataset_headers = ["id", "name", "categories"]
-
-        book_resource = BookResource()
-        book_resource._meta.skip_unchanged = True
-
-        # import with natural order
-        dataset_row = [book.id, book.name, f"{cat1.id},{cat2.id}"]
-        dataset = tablib.Dataset(headers=dataset_headers)
-        dataset.append(dataset_row)
-
-        result = book_resource.import_data(dataset, dry_run=False)
-        self.assertEqual(result.rows[0].import_type, results.RowResult.IMPORT_TYPE_SKIP)
-
-        # import with reverse order
-        dataset_row = [book.id, book.name, f"{cat2.id},{cat1.id}"]
-        dataset = tablib.Dataset(headers=dataset_headers)
-        dataset.append(dataset_row)
-
-        result = book_resource.import_data(dataset, dry_run=False)
-        self.assertEqual(result.rows[0].import_type, results.RowResult.IMPORT_TYPE_SKIP)
-
-        self.assertEqual(2, book.categories.count())
-
-    @ignore_widget_deprecation_warning
-    def test_many_to_many_widget_handles_uuid(self):
-        # Test for #1435 - skip_row() handles M2M field when UUID pk used
-        class _UUIDBookResource(resources.ModelResource):
-            class Meta:
-                model = UUIDBook
-
-        uuid_resource = _UUIDBookResource()
-        uuid_resource._meta.skip_unchanged = True
-        cat1 = UUIDCategory.objects.create(name="Category 1")
-        cat2 = UUIDCategory.objects.create(name="Category 2")
-        uuid_book = UUIDBook.objects.create(name="uuid book")
-        uuid_book.categories.add(cat1, cat2)
-        uuid_book.save()
-
-        dataset_headers = ["id", "name", "categories"]
-        dataset_row = [uuid_book.id, uuid_book.name, f"{cat1.catid},{cat2.catid}"]
-        dataset = tablib.Dataset(headers=dataset_headers)
-        dataset.append(dataset_row)
-        result = uuid_resource.import_data(dataset, dry_run=False)
-        self.assertEqual(result.rows[0].import_type, results.RowResult.IMPORT_TYPE_SKIP)
-
-    @ignore_widget_deprecation_warning
-    def test_skip_row_no_m2m_data_supplied(self):
-        # issue #1437
-        # test skip_row() when the model defines a m2m field
-        # but it is not present in the dataset
-        book = Book.objects.first()
-        dataset_headers = ["id", "name"]
-        dataset_row = [book.id, book.name]
-        dataset = tablib.Dataset(headers=dataset_headers)
-        dataset.append(dataset_row)
-
-        book_resource = BookResource()
-        book_resource._meta.skip_unchanged = True
-
-        self.assertEqual(1, book.categories.count())
-        result = book_resource.import_data(dataset, dry_run=False)
-        self.assertEqual(result.rows[0].import_type, results.RowResult.IMPORT_TYPE_SKIP)
-        self.assertEqual(1, book.categories.count())
-
-
-@mock.patch("import_export.resources.Diff", spec=True)
-class SkipDiffTest(TestCase):
-    """
-    Tests that the meta attribute 'skip_diff' means that no diff operations are called.
-    'copy.deepcopy' cannot be patched at class level because it causes interferes with
-    ``resources.Resource.__init__()``.
-    """
-
-    def setUp(self):
-        class _BookResource(resources.ModelResource):
-            class Meta:
-                model = Book
-                skip_diff = True
-
-        self.resource = _BookResource()
-        self.dataset = tablib.Dataset(headers=["id", "name", "birthday"])
-        self.dataset.append(["", "A.A.Milne", "1882test-01-18"])
-
-    def test_skip_diff(self, mock_diff):
-        with mock.patch("import_export.resources.deepcopy") as mock_deep_copy:
-            self.resource.import_data(self.dataset)
-            mock_diff.return_value.compare_with.assert_not_called()
-            mock_diff.return_value.as_html.assert_not_called()
-            mock_deep_copy.assert_not_called()
-
-    def test_skip_diff_for_delete_new_resource(self, mock_diff):
-        class BookResource(resources.ModelResource):
-            class Meta:
-                model = Book
-                skip_diff = True
-
-            def for_delete(self, row, instance):
-                return True
-
-        resource = BookResource()
-        with mock.patch("import_export.resources.deepcopy") as mock_deep_copy:
-            resource.import_data(self.dataset)
-            mock_diff.return_value.compare_with.assert_not_called()
-            mock_diff.return_value.as_html.assert_not_called()
-            mock_deep_copy.assert_not_called()
-
-    def test_skip_diff_for_delete_existing_resource(self, mock_diff):
-        book = Book.objects.create()
-
-        class BookResource(resources.ModelResource):
-            class Meta:
-                model = Book
-                skip_diff = True
-
-            def get_or_init_instance(self, instance_loader, row):
-                return book, False
-
-            def for_delete(self, row, instance):
-                return True
-
-        resource = BookResource()
-
-        with mock.patch("import_export.resources.deepcopy") as mock_deep_copy:
-            resource.import_data(self.dataset, dry_run=True)
-            mock_diff.return_value.compare_with.assert_not_called()
-            mock_diff.return_value.as_html.assert_not_called()
-            mock_deep_copy.assert_not_called()
-
-    def test_skip_diff_for_delete_skip_row_not_enabled_new_object(self, mock_diff):
-        class BookResource(resources.ModelResource):
-            class Meta:
-                model = Book
-                skip_diff = False
-
-            def for_delete(self, row, instance):
-                return True
-
-        resource = BookResource()
-
-        with mock.patch("import_export.resources.deepcopy") as mock_deep_copy:
-            resource.import_data(self.dataset, dry_run=True)
-            self.assertEqual(1, mock_diff.return_value.compare_with.call_count)
-            self.assertEqual(1, mock_deep_copy.call_count)
-
-    def test_skip_row_returns_false_when_skip_diff_is_true(self, mock_diff):
-        class BookResource(resources.ModelResource):
-            class Meta:
-                model = Book
-                skip_unchanged = True
-                skip_diff = True
-
-        resource = BookResource()
-
-        with mock.patch(
-            "import_export.resources.Resource.get_import_fields"
-        ) as mock_get_import_fields:
-            resource.import_data(self.dataset, dry_run=True)
-            self.assertEqual(2, mock_get_import_fields.call_count)
-
-
-class SkipHtmlDiffTest(TestCase):
-    def test_skip_html_diff(self):
-        class BookResource(resources.ModelResource):
-            class Meta:
-                model = Book
-                skip_html_diff = True
-
-        resource = BookResource()
-        self.dataset = tablib.Dataset(headers=["id", "name", "birthday"])
-        self.dataset.append(["", "A.A.Milne", "1882test-01-18"])
-
-        with mock.patch("import_export.resources.Diff.as_html") as mock_as_html:
-            resource.import_data(self.dataset, dry_run=True)
-            mock_as_html.assert_not_called()
-
-
-class BulkTest(TestCase):
-    def setUp(self):
-        class _BookResource(resources.ModelResource):
-            class Meta:
-                model = Book
-                use_bulk = True
-
-        self.resource = _BookResource()
-        rows = [(i + 1, "book_name") for i in range(10)]
-        self.dataset = tablib.Dataset(*rows, headers=["id", "name"])
-
-    def init_update_test_data(self, model=Book):
-        [model.objects.create(name="book_name") for _ in range(10)]
-        self.assertEqual(10, model.objects.count())
-        rows = model.objects.all().values_list("id", "name")
-        updated_rows = [(r[0], "UPDATED") for r in rows]
-        self.dataset = tablib.Dataset(*updated_rows, headers=["id", "name"])
-
-
-class BulkCreateTest(BulkTest):
-    @mock.patch("core.models.Book.objects.bulk_create")
-    @ignore_widget_deprecation_warning
-    def test_bulk_create_does_not_call_object_save(self, mock_bulk_create):
-        with mock.patch("core.models.Book.save") as mock_obj_save:
-            self.resource.import_data(self.dataset)
-            mock_obj_save.assert_not_called()
-        mock_bulk_create.assert_called_with(mock.ANY, batch_size=None)
-
-    @mock.patch("core.models.Book.objects.bulk_create")
-    @ignore_widget_deprecation_warning
-    def test_bulk_create_batch_size_of_5(self, mock_bulk_create):
-        class _BookResource(resources.ModelResource):
-            class Meta:
-                model = Book
-                use_bulk = True
-                batch_size = 5
-
-        resource = _BookResource()
-        result = resource.import_data(self.dataset)
-        self.assertEqual(2, mock_bulk_create.call_count)
-        mock_bulk_create.assert_called_with(mock.ANY, batch_size=5)
-        self.assertEqual(10, result.total_rows)
-
-    @mock.patch("core.models.UUIDBook.objects.bulk_create")
-    @ignore_widget_deprecation_warning
-    def test_bulk_create_uuid_model(self, mock_bulk_create):
-        """Test create of a Model which defines uuid not pk (issue #1274)"""
-
-        class _UUIDBookResource(resources.ModelResource):
-            class Meta:
-                model = UUIDBook
-                use_bulk = True
-                batch_size = 5
-                fields = (
-                    "id",
-                    "name",
-                )
-
-        resource = _UUIDBookResource()
-        result = resource.import_data(self.dataset)
-        self.assertEqual(2, mock_bulk_create.call_count)
-        mock_bulk_create.assert_called_with(mock.ANY, batch_size=5)
-        self.assertEqual(10, result.total_rows)
-
-    @mock.patch("core.models.Book.objects.bulk_create")
-    @ignore_widget_deprecation_warning
-    def test_bulk_create_no_batch_size(self, mock_bulk_create):
-        class _BookResource(resources.ModelResource):
-            class Meta:
-                model = Book
-                use_bulk = True
-                batch_size = None
-
-        resource = _BookResource()
-        result = resource.import_data(self.dataset)
-        self.assertEqual(1, mock_bulk_create.call_count)
-        mock_bulk_create.assert_called_with(mock.ANY, batch_size=None)
-        self.assertEqual(10, result.total_rows)
-        self.assertEqual(10, result.totals["new"])
-
-    @mock.patch("core.models.Book.objects.bulk_create")
-    @ignore_widget_deprecation_warning
-    def test_bulk_create_called_dry_run(self, mock_bulk_create):
-        class _BookResource(resources.ModelResource):
-            class Meta:
-                model = Book
-                use_bulk = True
-                batch_size = None
-
-        resource = _BookResource()
-        result = resource.import_data(self.dataset, dry_run=True)
-        self.assertEqual(1, mock_bulk_create.call_count)
-        self.assertEqual(10, result.total_rows)
-        self.assertEqual(10, result.totals["new"])
-
-    @mock.patch("core.models.Book.objects.bulk_create")
-    @ignore_widget_deprecation_warning
-    def test_bulk_create_not_called_when_not_using_transactions(self, mock_bulk_create):
-        class _BookResource(resources.ModelResource):
-            def import_data(
-                self,
-                dataset,
-                dry_run=False,
-                raise_errors=False,
-                use_transactions=None,
-                collect_failed_rows=False,
-                **kwargs,
-            ):
-                # override so that we can enforce not using_transactions
-                using_transactions = False
-                return self.import_data_inner(
-                    dataset,
-                    dry_run,
-                    raise_errors,
-                    using_transactions,
-                    collect_failed_rows,
-                    **kwargs,
-                )
-
-            class Meta:
-                model = Book
-                use_bulk = True
-
-        resource = _BookResource()
-        resource.import_data(self.dataset, dry_run=True)
-        mock_bulk_create.assert_not_called()
-
-    @mock.patch("core.models.Book.objects.bulk_create")
-    @ignore_widget_deprecation_warning
-    def test_bulk_create_batch_size_of_4(self, mock_bulk_create):
-        class _BookResource(resources.ModelResource):
-            class Meta:
-                model = Book
-                use_bulk = True
-                batch_size = 4
-
-        resource = _BookResource()
-        result = resource.import_data(self.dataset)
-        self.assertEqual(3, mock_bulk_create.call_count)
-        self.assertEqual(10, result.total_rows)
-        self.assertEqual(10, result.totals["new"])
-
-    @ignore_widget_deprecation_warning
-    def test_no_changes_for_errors_if_use_transactions_enabled(self):
-        with mock.patch("import_export.results.Result.has_errors") as mock_has_errors:
-            mock_has_errors.return_val = True
-            self.resource.import_data(self.dataset)
-        self.assertEqual(0, Book.objects.count())
-
-    @mock.patch("core.models.Book.objects.bulk_create")
-    @ignore_widget_deprecation_warning
-    def test_bulk_create_use_bulk_disabled(self, mock_bulk_create):
-        class _BookResource(resources.ModelResource):
-            class Meta:
-                model = Book
-                use_bulk = False
-
-        resource = _BookResource()
-        result = resource.import_data(self.dataset)
-        mock_bulk_create.assert_not_called()
-        self.assertEqual(10, Book.objects.count())
-        self.assertEqual(10, result.total_rows)
-        self.assertEqual(10, result.totals["new"])
-
-    @mock.patch("core.models.Book.objects.bulk_create")
-    @ignore_widget_deprecation_warning
-    def test_bulk_create_bad_batch_size_value(self, mock_bulk_create):
-        class _BookResource(resources.ModelResource):
-            class Meta:
-                model = Book
-                use_bulk = True
-                batch_size = "a"
-
-        resource = _BookResource()
-        with self.assertRaises(ValueError):
-            resource.import_data(self.dataset)
-        mock_bulk_create.assert_not_called()
-
-    @mock.patch("core.models.Book.objects.bulk_create")
-    def test_bulk_create_negative_batch_size_value(self, mock_bulk_create):
-        class _BookResource(resources.ModelResource):
-            class Meta:
-                model = Book
-                use_bulk = True
-                batch_size = -1
-
-        resource = _BookResource()
-        with self.assertRaises(ValueError):
-            resource.import_data(self.dataset)
-        mock_bulk_create.assert_not_called()
-
-    @mock.patch("core.models.Book.objects.bulk_create")
-    @ignore_widget_deprecation_warning
-    def test_bulk_create_oversized_batch_size_value(self, mock_bulk_create):
-        class _BookResource(resources.ModelResource):
-            class Meta:
-                model = Book
-                use_bulk = True
-                batch_size = 100
-
-        resource = _BookResource()
-        result = resource.import_data(self.dataset)
-        self.assertEqual(1, mock_bulk_create.call_count)
-        mock_bulk_create.assert_called_with(mock.ANY, batch_size=None)
-        self.assertEqual(10, result.total_rows)
-        self.assertEqual(10, result.totals["new"])
-
-    @mock.patch("core.models.Book.objects.bulk_create")
-    @ignore_widget_deprecation_warning
-    def test_bulk_create_logs_exception(self, mock_bulk_create):
-        e = ValidationError("invalid field")
-        mock_bulk_create.side_effect = e
-
-        class _BookResource(resources.ModelResource):
-            class Meta:
-                model = Book
-                use_bulk = True
-                batch_size = 100
-
-        resource = _BookResource()
-        with mock.patch("logging.Logger.debug") as mock_exception:
-            resource.import_data(self.dataset)
-            mock_exception.assert_called_with(e, exc_info=e)
-
-    @mock.patch("core.models.Book.objects.bulk_create")
-    @ignore_widget_deprecation_warning
-    def test_bulk_create_raises_exception(self, mock_bulk_create):
-        mock_bulk_create.side_effect = ValidationError("invalid field")
-
-        class _BookResource(resources.ModelResource):
-            class Meta:
-                model = Book
-                use_bulk = True
-                batch_size = 100
-
-        resource = _BookResource()
-        with self.assertRaises(ValidationError):
-            resource.import_data(self.dataset, raise_errors=True)
-
-    @mock.patch("core.models.Book.objects.bulk_create")
-    @ignore_widget_deprecation_warning
-    def test_bulk_create_exception_gathered_on_dry_run(self, mock_bulk_create):
-        mock_bulk_create.side_effect = ValidationError("invalid field")
-
-        class _BookResource(resources.ModelResource):
-            class Meta:
-                model = Book
-                use_bulk = True
-                batch_size = 100
-
-        resource = _BookResource()
-        result = resource.import_data(self.dataset, dry_run=True, raise_errors=False)
-        self.assertTrue(result.has_errors())
-
-    @ignore_widget_deprecation_warning
-    def test_m2m_not_called_for_bulk(self):
-        mock_m2m_widget = mock.Mock(spec=widgets.ManyToManyWidget)
-
-        class BookM2MResource(resources.ModelResource):
-            categories = fields.Field(attribute="categories", widget=mock_m2m_widget)
-
-            class Meta:
-                model = Book
-                use_bulk = True
-
-        resource = BookM2MResource()
-        self.dataset.append_col(["Cat 1|Cat 2"] * 10, header="categories")
-        resource.import_data(self.dataset, raise_errors=True)
-        mock_m2m_widget.assert_not_called()
-
-    def test_force_init_instance(self):
-        class _BookResource(resources.ModelResource):
-            def get_instance(self, instance_loader, row):
-                raise AssertionError("should not be called")
-
-            class Meta:
-                model = Book
-                force_init_instance = True
-
-        resource = _BookResource()
-        self.assertIsNotNone(
-            resource.get_or_init_instance(
-                ModelInstanceLoader(resource), self.dataset[0]
-            )
-        )
-
-    @mock.patch("import_export.resources.atomic_if_using_transaction")
-    def test_no_sub_transaction_on_row_for_bulk(self, mock_atomic_if_using_transaction):
-        class _BookResource(resources.ModelResource):
-            class Meta:
-                model = Book
-                use_bulk = True
-
-        resource = _BookResource()
-        resource.import_data(self.dataset)
-        self.assertIn(
-            False, [x[0][0] for x in mock_atomic_if_using_transaction.call_args_list]
-        )
-
-
-class BulkUpdateTest(BulkTest):
-    class _BookResource(resources.ModelResource):
-        class Meta:
-            model = Book
-            use_bulk = True
-            fields = ("id", "name")
-            import_id_fields = ("id",)
-
-    def setUp(self):
-        super().setUp()
-        self.init_update_test_data()
-        self.resource = self._BookResource()
-
-    @ignore_widget_deprecation_warning
-    def test_bulk_update(self):
-        result = self.resource.import_data(self.dataset)
-        [self.assertEqual("UPDATED", b.name) for b in Book.objects.all()]
-        self.assertEqual(10, result.total_rows)
-        self.assertEqual(10, result.totals["update"])
-
-    @mock.patch("core.models.Book.objects.bulk_update")
-    @ignore_widget_deprecation_warning
-    def test_bulk_update_batch_size_of_4(self, mock_bulk_update):
-        class _BookResource(resources.ModelResource):
-            class Meta:
-                model = Book
-                use_bulk = True
-                batch_size = 4
-
-        resource = _BookResource()
-        result = resource.import_data(self.dataset)
-        self.assertEqual(3, mock_bulk_update.call_count)
-        self.assertEqual(10, result.total_rows)
-        self.assertEqual(10, result.totals["update"])
-
-    @mock.patch("core.models.Book.objects.bulk_update")
-    @ignore_widget_deprecation_warning
-    def test_bulk_update_batch_size_of_5(self, mock_bulk_update):
-        class _BookResource(resources.ModelResource):
-            class Meta:
-                model = Book
-                use_bulk = True
-                batch_size = 5
-
-        resource = _BookResource()
-        result = resource.import_data(self.dataset)
-        self.assertEqual(2, mock_bulk_update.call_count)
-        self.assertEqual(10, result.total_rows)
-        self.assertEqual(10, result.totals["update"])
-
-    @mock.patch("core.models.Book.objects.bulk_update")
-    @ignore_widget_deprecation_warning
-    def test_bulk_update_no_batch_size(self, mock_bulk_update):
-        class _BookResource(resources.ModelResource):
-            class Meta:
-                model = Book
-                use_bulk = True
-                batch_size = None
-
-        resource = _BookResource()
-        result = resource.import_data(self.dataset)
-        self.assertEqual(1, mock_bulk_update.call_count)
-        mock_bulk_update.assert_called_with(mock.ANY, mock.ANY, batch_size=None)
-        self.assertEqual(10, result.total_rows)
-        self.assertEqual(10, result.totals["update"])
-
-    @mock.patch("core.models.Book.objects.bulk_update")
-    @ignore_widget_deprecation_warning
-    def test_bulk_update_not_called_when_not_using_transactions(self, mock_bulk_update):
-        class _BookResource(resources.ModelResource):
-            def import_data(
-                self,
-                dataset,
-                dry_run=False,
-                raise_errors=False,
-                use_transactions=None,
-                collect_failed_rows=False,
-                **kwargs,
-            ):
-                # override so that we can enforce not using_transactions
-                using_transactions = False
-                return self.import_data_inner(
-                    dataset,
-                    dry_run,
-                    raise_errors,
-                    using_transactions,
-                    collect_failed_rows,
-                    **kwargs,
-                )
-
-            class Meta:
-                model = Book
-                use_bulk = True
-
-        resource = _BookResource()
-        resource.import_data(self.dataset, dry_run=True)
-        mock_bulk_update.assert_not_called()
-
-    @mock.patch("core.models.Book.objects.bulk_update")
-    @ignore_widget_deprecation_warning
-    def test_bulk_update_called_for_dry_run(self, mock_bulk_update):
-        self.resource.import_data(self.dataset, dry_run=True)
-        self.assertEqual(1, mock_bulk_update.call_count)
-
-    @mock.patch("core.models.Book.objects.bulk_update")
-    @ignore_widget_deprecation_warning
-    def test_bulk_not_called_when_use_bulk_disabled(self, mock_bulk_update):
-        class _BookResource(resources.ModelResource):
-            class Meta:
-                model = Book
-                use_bulk = False
-
-        resource = _BookResource()
-        result = resource.import_data(self.dataset)
-        self.assertEqual(10, Book.objects.count())
-        self.assertEqual(10, result.total_rows)
-        self.assertEqual(10, result.totals["update"])
-        mock_bulk_update.assert_not_called()
-
-    @mock.patch("core.models.Book.objects.bulk_update")
-    @ignore_widget_deprecation_warning
-    def test_bulk_update_logs_exception(self, mock_bulk_update):
-        e = ValidationError("invalid field")
-        mock_bulk_update.side_effect = e
-
-        class _BookResource(resources.ModelResource):
-            class Meta:
-                model = Book
-                use_bulk = True
-
-        resource = _BookResource()
-        with mock.patch("logging.Logger.debug") as mock_exception:
-            resource.import_data(self.dataset)
-            mock_exception.assert_called_with(e, exc_info=e)
-
-    @mock.patch("core.models.Book.objects.bulk_update")
-    @ignore_widget_deprecation_warning
-    def test_bulk_update_raises_exception(self, mock_bulk_update):
-        e = ValidationError("invalid field")
-        mock_bulk_update.side_effect = e
-
-        class _BookResource(resources.ModelResource):
-            class Meta:
-                model = Book
-                use_bulk = True
-
-        resource = _BookResource()
-        with self.assertRaises(ValidationError) as raised_exc:
-            resource.import_data(self.dataset, raise_errors=True)
-            self.assertEqual(e, raised_exc)
-
-
-class BulkUUIDBookUpdateTest(BulkTest):
-    def setUp(self):
-        super().setUp()
-        self.init_update_test_data(model=UUIDBook)
-
-    @mock.patch("core.models.UUIDBook.objects.bulk_update")
-    @ignore_widget_deprecation_warning
-    def test_bulk_update_uuid_model(self, mock_bulk_update):
-        """Test update of a Model which defines uuid not pk (issue #1274)"""
-
-        class _UUIDBookResource(resources.ModelResource):
-            class Meta:
-                model = UUIDBook
-                use_bulk = True
-                batch_size = 5
-                fields = (
-                    "id",
-                    "name",
-                )
-
-        resource = _UUIDBookResource()
-        result = resource.import_data(self.dataset)
-        self.assertEqual(2, mock_bulk_update.call_count)
-        self.assertEqual(10, result.total_rows)
-        self.assertEqual(10, result.totals["update"])
-
-
-class BulkDeleteTest(BulkTest):
-    class DeleteBookResource(resources.ModelResource):
-        def for_delete(self, row, instance):
-            return True
-
-        class Meta:
-            model = Book
-            use_bulk = True
-            # there are errors when diffing with mocks
-            # therefore disable diff with this flag
-            skip_diff = True
-
-    def setUp(self):
-        super().setUp()
-        self.resource = self.DeleteBookResource()
-        self.resource._meta.batch_size = 1000
-        self.resource._meta.use_bulk = True
-        self.init_update_test_data()
-
-    @mock.patch("core.models.Book.delete")
-    @ignore_widget_deprecation_warning
-    def test_bulk_delete_use_bulk_is_false(self, mock_obj_delete):
-        self.resource._meta.use_bulk = False
-        self.resource.import_data(self.dataset)
-        self.assertEqual(10, mock_obj_delete.call_count)
-
-    @mock.patch("core.models.Book.objects")
-    @ignore_widget_deprecation_warning
-    def test_bulk_delete_batch_size_of_4(self, mock_obj_manager):
-        self.resource._meta.batch_size = 4
-        result = self.resource.import_data(self.dataset)
-        self.assertEqual(3, mock_obj_manager.filter.return_value.delete.call_count)
-        self.assertEqual(10, result.total_rows)
-        self.assertEqual(10, result.totals["delete"])
-
-    @mock.patch("core.models.Book.objects")
-    @ignore_widget_deprecation_warning
-    def test_bulk_delete_batch_size_of_5(self, mock_obj_manager):
-        self.resource._meta.batch_size = 5
-        result = self.resource.import_data(self.dataset)
-        self.assertEqual(2, mock_obj_manager.filter.return_value.delete.call_count)
-        self.assertEqual(10, result.total_rows)
-        self.assertEqual(10, result.totals["delete"])
-
-    @mock.patch("core.models.Book.objects")
-    @ignore_widget_deprecation_warning
-    def test_bulk_delete_batch_size_is_none(self, mock_obj_manager):
-        self.resource._meta.batch_size = None
-        result = self.resource.import_data(self.dataset)
-        self.assertEqual(1, mock_obj_manager.filter.return_value.delete.call_count)
-        self.assertEqual(10, result.total_rows)
-        self.assertEqual(10, result.totals["delete"])
-
-    @mock.patch("core.models.Book.objects")
-    @ignore_widget_deprecation_warning
-    def test_bulk_delete_not_called_when_not_using_transactions(self, mock_obj_manager):
-        class _BookResource(self.DeleteBookResource):
-            def import_data(
-                self,
-                dataset,
-                dry_run=False,
-                raise_errors=False,
-                use_transactions=None,
-                collect_failed_rows=False,
-                **kwargs,
-            ):
-                # override so that we can enforce not using_transactions
-                using_transactions = False
-                return self.import_data_inner(
-                    dataset,
-                    dry_run,
-                    raise_errors,
-                    using_transactions,
-                    collect_failed_rows,
-                    **kwargs,
-                )
-
-        resource = _BookResource()
-        resource.import_data(self.dataset, dry_run=True)
-        self.assertEqual(0, mock_obj_manager.filter.return_value.delete.call_count)
-
-    @mock.patch("core.models.Book.objects")
-    @ignore_widget_deprecation_warning
-    def test_bulk_delete_called_for_dry_run(self, mock_obj_manager):
-        self.resource.import_data(self.dataset, dry_run=True)
-        self.assertEqual(1, mock_obj_manager.filter.return_value.delete.call_count)
-
-    @mock.patch("core.models.Book.objects")
-    @ignore_widget_deprecation_warning
-    def test_bulk_delete_logs_exception(self, mock_obj_manager):
-        e = Exception("invalid")
-        mock_obj_manager.filter.return_value.delete.side_effect = e
-
-        with mock.patch("logging.Logger.debug") as mock_exception:
-            self.resource.import_data(self.dataset)
-            mock_exception.assert_called_with(e, exc_info=mock.ANY)
-            self.assertEqual(1, mock_exception.call_count)
-
-    @mock.patch("core.models.Book.objects")
-    @ignore_widget_deprecation_warning
-    def test_bulk_delete_raises_exception(self, mock_obj_manager):
-        e = Exception("invalid")
-        mock_obj_manager.filter.return_value.delete.side_effect = e
-
-        with self.assertRaises(Exception) as raised_exc:
-            self.resource.import_data(self.dataset, raise_errors=True)
-            self.assertEqual(e, raised_exc)
-
-
-class BulkUUIDBookDeleteTest(BulkTest):
-    class DeleteBookResource(resources.ModelResource):
-        def for_delete(self, row, instance):
-            return True
-
-        class Meta:
-            model = UUIDBook
-            use_bulk = True
-            batch_size = 5
-
-    def setUp(self):
-        super().setUp()
-        self.resource = self.DeleteBookResource()
-        self.init_update_test_data(model=UUIDBook)
-
-    @ignore_widget_deprecation_warning
-    def test_bulk_delete_batch_size_of_5(self):
-        self.assertEqual(10, UUIDBook.objects.count())
-        self.resource.import_data(self.dataset)
-        self.assertEqual(0, UUIDBook.objects.count())
-
-
-class RawValueTest(TestCase):
-    def setUp(self):
-        class _BookResource(resources.ModelResource):
-            class Meta:
-                model = Book
-                store_row_values = True
-
-        self.resource = _BookResource()
-
-        self.book = Book.objects.create(name="Some book")
-        self.dataset = tablib.Dataset(headers=["id", "name", "author_email", "price"])
-        row = [self.book.pk, "Some book", "test@example.com", "10.25"]
-        self.dataset.append(row)
-
-    @ignore_widget_deprecation_warning
-    def test_import_data(self):
-        result = self.resource.import_data(self.dataset, raise_errors=True)
-
-        self.assertFalse(result.has_errors())
-        self.assertEqual(len(result.rows), 1)
-        self.assertTrue(result.rows[0].diff)
-        self.assertEqual(
-            result.rows[0].import_type, results.RowResult.IMPORT_TYPE_UPDATE
-        )
-        self.assertEqual(result.rows[0].row_values.get("name"), "Some book")
-        self.assertEqual(
-            result.rows[0].row_values.get("author_email"), "test@example.com"
-        )
-        self.assertEqual(result.rows[0].row_values.get("price"), "10.25")
-
-
-class ResourcesHelperFunctionsTest(TestCase):
-    """
-    Test the helper functions in resources.
-    """
-
-    def test_has_natural_foreign_key(self):
-        """
-        Ensure that resources.has_natural_foreign_key detects correctly
-        whether a model has a natural foreign key
-        """
-        cases = {Book: True, Author: True, Category: False}
-
-        for model, expected_result in cases.items():
-            self.assertEqual(resources.has_natural_foreign_key(model), expected_result)
-
-
-class AfterImportComparisonTest(TestCase):
-    class BookResource(resources.ModelResource):
-        is_published = False
-
-        def after_import_row(self, row, row_result, **kwargs):
-            if (
-                getattr(row_result.original, "published") is None
-                and getattr(row_result.instance, "published") is not None
-            ):
-                self.is_published = True
-
-        class Meta:
-            model = Book
-            store_instance = True
-
-    def setUp(self):
-        super().setUp()
-        self.resource = AfterImportComparisonTest.BookResource()
-        self.book = Book.objects.create(name="Some book")
-        self.dataset = tablib.Dataset(headers=["id", "name", "published"])
-        row = [self.book.pk, "Some book", "2023-05-09"]
-        self.dataset.append(row)
-
-    @ignore_widget_deprecation_warning
-    def test_after_import_row_check_for_change(self):
-        # issue 1583 - assert that `original` object is available to after_import_row()
-        self.resource.import_data(self.dataset, raise_errors=True)
-        self.assertTrue(self.resource.is_published)
-
-
-class ImportExportFieldOrderTest(TestCase):
-    class BaseBookResource(resources.ModelResource):
-        def __init__(self):
-            self.field_names = list()
-
-        def get_queryset(self):
-            return Book.objects.all().order_by("id")
-
-        def import_field(self, field, obj, data, is_m2m=False, **kwargs):
-            # mock out import_field() so that we can see the order
-            # fields were called
-            self.field_names.append(field.column_name)
-
-    class UnorderedBookResource(BaseBookResource):
-        class Meta:
-            fields = ("price", "id", "name")
-            model = Book
-
-    class OrderedBookResource(BaseBookResource):
-        class Meta:
-            fields = ("price", "id", "name")
-            import_order = ("price", "name", "id")
-            export_order = ("price", "name", "id")
-            model = Book
-
-    class SubsetOrderedBookResource(BaseBookResource):
-        class Meta:
-            fields = ("price", "id", "name", "published")
-            import_order = ("name",)
-            export_order = ("published",)
-            model = Book
-
-    class DuplicateFieldsBookResource(BaseBookResource):
-        class Meta:
-            fields = ("id", "price", "name", "price")
-            model = Book
-
-    class FieldsAsListBookResource(BaseBookResource):
-        class Meta:
-            fields = ["id", "price", "name"]
-            model = Book
-
-    def setUp(self):
-        super().setUp()
-        self.pk = Book.objects.create(name="Ulysses", price="1.99").pk
-        self.dataset = tablib.Dataset(headers=["id", "name", "price"])
-        row = [self.pk, "Some book", "19.99"]
-        self.dataset.append(row)
-
-    @ignore_widget_deprecation_warning
-    def test_defined_import_order(self):
-        self.resource = ImportExportFieldOrderTest.OrderedBookResource()
-        self.resource.import_data(self.dataset)
-        self.assertEqual(["price", "name", "id"], self.resource.field_names)
-
-    @ignore_widget_deprecation_warning
-    def test_undefined_import_order(self):
-        self.resource = ImportExportFieldOrderTest.UnorderedBookResource()
-        self.resource.import_data(self.dataset)
-        self.assertEqual(["price", "id", "name"], self.resource.field_names)
-
-    @ignore_widget_deprecation_warning
-    def test_defined_export_order(self):
-        self.resource = ImportExportFieldOrderTest.OrderedBookResource()
-        data = self.resource.export()
-        target = f"price,name,id\r\n1.99,Ulysses,{self.pk}\r\n"
-        self.assertEqual(target, data.csv)
-
-    @ignore_widget_deprecation_warning
-    def test_undefined_export_order(self):
-        # When export order is not defined,
-        # exported order should correspond with 'fields' definition
-        self.resource = ImportExportFieldOrderTest.UnorderedBookResource()
-        data = self.resource.export()
-        target = f"price,id,name\r\n1.99,{self.pk},Ulysses\r\n"
-        self.assertEqual(target, data.csv)
-
-    @ignore_widget_deprecation_warning
-    def test_subset_import_order(self):
-        self.resource = ImportExportFieldOrderTest.SubsetOrderedBookResource()
-        self.resource.import_data(self.dataset)
-        self.assertEqual(
-            ["name", "price", "id", "published"], self.resource.field_names
-        )
-
-    @ignore_widget_deprecation_warning
-    def test_subset_export_order(self):
-        self.resource = ImportExportFieldOrderTest.SubsetOrderedBookResource()
-        data = self.resource.export()
-        target = f"published,price,id,name\r\n,1.99,{self.pk},Ulysses\r\n"
-        self.assertEqual(target, data.csv)
-
-    @ignore_widget_deprecation_warning
-    def test_duplicate_import_order(self):
-        self.resource = ImportExportFieldOrderTest.DuplicateFieldsBookResource()
-        self.resource.import_data(self.dataset)
-        self.assertEqual(["id", "price", "name"], self.resource.field_names)
-
-    @ignore_widget_deprecation_warning
-    def test_duplicate_export_order(self):
-        self.resource = ImportExportFieldOrderTest.DuplicateFieldsBookResource()
-        data = self.resource.export()
-        target = f"id,price,name\r\n{self.pk},1.99,Ulysses\r\n"
-        self.assertEqual(target, data.csv)
-
-    @ignore_widget_deprecation_warning
-    def test_fields_as_list_import_order(self):
-        self.resource = ImportExportFieldOrderTest.FieldsAsListBookResource()
-        self.resource.import_data(self.dataset)
-        self.assertEqual(["id", "price", "name"], self.resource.field_names)
-
-    @ignore_widget_deprecation_warning
-    def test_fields_as_list_export_order(self):
-        self.resource = ImportExportFieldOrderTest.FieldsAsListBookResource()
-        data = self.resource.export()
-        target = f"id,price,name\r\n{self.pk},1.99,Ulysses\r\n"
-        self.assertEqual(target, data.csv)
-
-
-class ImportIdFieldsTestCase(TestCase):
-    class BookResource(resources.ModelResource):
-        name = fields.Field(attribute="name", column_name="book_name")
-
-        class Meta:
-            model = Book
-            import_id_fields = ["name"]
-
-    def setUp(self):
-        super().setUp()
-        self.book = Book.objects.create(name="The Hobbit")
-        self.resource = ImportIdFieldsTestCase.BookResource()
-
-    def test_custom_column_name_warns_if_not_present(self):
-        dataset = tablib.Dataset(
-            *[(self.book.pk, "Some book")], headers=["id", "wrong_name"]
-        )
-        with self.assertRaisesRegex(
-            exceptions.FieldError,
-            "The following import_id_fields are not present "
-            "in the dataset: book_name",
-        ):
-            self.resource.import_data(dataset)
-
-    def test_multiple_import_id_fields(self):
-        class BookResource(resources.ModelResource):
-            class Meta:
-                model = Book
-                import_id_fields = ("id", "name", "author_email")
-
-        self.resource = BookResource()
-        dataset = tablib.Dataset(
-            *[(self.book.pk, "Goldeneye", "ian.fleming@example.com")],
-            headers=["A", "name", "B"],
-        )
-        with self.assertRaisesRegex(
-            exceptions.FieldError,
-            "The following import_id_fields are not present "
-            "in the dataset: id, author_email",
-        ):
-            self.resource.import_data(dataset)
-
-
-class ImportWithMissingFields(TestCase):
-    # issue 1517
-
-    @patch("import_export.resources.logger")
-    @patch("import_export.fields.Field.save")
-    @ignore_widget_deprecation_warning
-    def test_import_with_missing_instance_attribute(self, mock_field_save, mock_logger):
-        class _BookResource(resources.ModelResource):
-            name = fields.Field(column_name="name")
-
-            class Meta:
-                model = Book
-
-        dataset = tablib.Dataset(*[(1, "Some book")], headers=["id", "name"])
-        self.resource = _BookResource()
-        result = self.resource.import_data(dataset)
-        self.assertFalse(result.has_errors())
-        target = (
-            "skipping field '<import_export.fields.Field: name>' "
-            "- field attribute is not defined"
-        )
-        mock_logger.debug.assert_any_call(target)
-        self.assertEqual(1, mock_field_save.call_count)
-
-    @patch("import_export.resources.logger")
-    @patch("import_export.fields.Field.save")
-    @ignore_widget_deprecation_warning
-    def test_import_with_missing_field_in_row(self, mock_field_save, mock_logger):
-        dataset = tablib.Dataset(*[(1, "Some book")], headers=["id", "name"])
-        self.resource = BookResource()
-        result = self.resource.import_data(dataset)
-        self.assertFalse(result.has_errors())
-        mock_logger.debug.assert_any_call(
-            "skipping field '<import_export.fields.Field: author_email>' "
-            "- column name 'author_email' is not present in row"
-        )
-        self.assertEqual(2, mock_field_save.call_count)
