+++ tests/core/tests/widget.py
@@ -1,0 +1,6 @@
+from import_export import widgets
+
+
+class HarshRussianWidget(widgets.CharWidget):
+    def clean(self, value, row=None, *args, **kwargs):
+        raise ValueError("Ова вриједност је страшна!")
